updated to use @qavajs/core (#145)
