+++ .github/workflows/pull-request.yml
@@ -4,6 +4,8 @@ on:
   pull_request:
     branches:
       - main
+      - v1
+      - v2
 
 jobs:
   unix:
