+++ src/actions.ts
@@ -1,6 +1,6 @@
 import { When } from '@cucumber/cucumber';
 import { parseCoords, parseKeySequence, parseCoordsAsObject, dragAndDrop } from './utils';
-import { MemoryValue } from '@qavajs/cli';
+import { MemoryValue } from '@qavajs/core';
 import { Locator } from './pageObject';
 
 /**
