+++ src/memory.ts
@@ -1,6 +1,6 @@
 import { When } from '@cucumber/cucumber';
 import { Locator } from './pageObject';
-import { MemoryValue } from '@qavajs/cli';
+import { MemoryValue } from '@qavajs/core';
 
 /**
  * Save text of element to memory
