+++ test/poImport.spec.ts
@@ -1,15 +1,9 @@
-import { po, $, $$, Component, Selector, NativeSelector } from '../po'
+import { locator } from '../po'
 import { test, expect } from 'vitest';
 test.each([
-    po.init,
-    po.register,
-    po.setDriver,
-    po.getElement,
-    $,
-    $$,
-    Component,
-    Selector,
-    NativeSelector
+    locator,
+    locator.template,
+    locator.native
 ])('po function %o', (fn) => {
     expect(fn).toBeInstanceOf(Function);
 });
