--- test/utilsImport.spec.ts
@@ -1,10 +1,0 @@
-import { getValue, getConditionWait, getLocator, getElement } from '../utils.js'
-import { test, expect } from 'vitest';
-test.each([
-    getValue,
-    getConditionWait,
-    getLocator,
-    getElement
-])('util function', (fn) => {
-    expect(fn).toBeInstanceOf(Function);
-});
