--- utils.js
@@ -1,6 +1,0 @@
-const { getValue, getElement, getLocator, getConditionWait } = require('./lib/transformers');
-
-exports.getValue = getValue;
-exports.getElement = getElement;
-exports.getLocator = getLocator;
-exports.getConditionWait = getConditionWait;
